+++ src/utils/db/getDocumentId.ts
